Fixed some warnings found by Studio (casts)
mix.c is now obsolete and eliminated from project.

git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@2079 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ srcphast/merge.c
@@ -465,19 +465,19 @@ void MergeBeginTimeStep(int print_sel, int print_out)
     /* Always open file for output in case of a warning message */
     {
       /* create the O.chem dataset */
-      FileInfo_dataset_create(&s_fiOutput, "O.chem", dims[0]);
+      FileInfo_dataset_create(&s_fiOutput, "O.chem", (int) dims[0]);
     } 
 
     if (print_sel == TRUE)
     {
       /* create the xyz.chem dataset */
-      FileInfo_dataset_create(&s_fiPunch, "xyz.chem", dims[0]);
+      FileInfo_dataset_create(&s_fiPunch, "xyz.chem", (int) dims[0]);
     }
 
     /* Always open file for echo in case of a warning message */
     {
       /* create the O.chem dataset */
-      FileInfo_dataset_create(&s_fiEcho, "log", dims[0]);
+      FileInfo_dataset_create(&s_fiEcho, "log", (int) dims[0]);
     } 
     /* init dspace coordinates */
     s_ci.coord[0][0] = 0;
